factor out `TranslateCF2short`
