+++ era_interim.py
@@ -50,7 +50,6 @@ from math                import exp, floor, atan2, pi
 from os                  import path, listdir, remove, makedirs
 from globsim.generic     import ParameterIO, StationListRead, ScaledFileOpen, series_interpolate, variables_skip, str_encode, cummulative2total, create_empty_netcdf, GenericDownload, GenericScale, GenericInterpolate
 from globsim.meteorology import spec_hum_kgkg, LW_downward, pressure_from_elevation
-from fnmatch             import filter
 from scipy.interpolate   import interp1d
 
 try:
@@ -835,6 +834,7 @@ class ERAIinterpolate(GenericInterpolate):
         
         """
         # open file 
+        
         ncf = nc.MFDataset(ncfile_in, 'r', aggdim='time')
         height = ncf.variables['height'][:]
         nt = len(ncf.variables['time'][:])
@@ -842,13 +842,8 @@ class ERAIinterpolate(GenericInterpolate):
         
         # list variables
         varlist = [str_encode(x) for x in ncf.variables.keys()]
-        varlist.remove('time')
-        varlist.remove('station')
-        varlist.remove('latitude')
-        varlist.remove('longitude')
-        varlist.remove('level')
-        varlist.remove('height')
-        varlist.remove('z')
+        for V in ['time', 'station', 'latitude', 'longitude', 'level','height','z']:
+            varlist.remove(V)
 
         # === open and prepare output netCDF file ==============================
         # dimensions: station, time
@@ -946,19 +941,6 @@ class ERAIinterpolate(GenericInterpolate):
         ncf.close()
         # closed file ==========================================================    
 
-
-    def TranslateCF2short(self, dpar):
-        """
-        Map CF Standard Names into short codes used in ERA-Interim netCDF files.
-        """
-        varlist = [] 
-        for var in self.variables:
-            varlist.append(dpar.get(var))
-        # drop none
-        varlist = [item for item in varlist if item is not None]      
-        # flatten
-        varlist = [item for sublist in varlist for item in sublist]         
-        return(varlist) 
     
     def process(self):
         """
